Updated cartpole structure.
